server/config/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt
djangorestframework-djangorestframework-jwt
django-cors-headers
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
